requirements.txt
discord.py
discord-py-slash-command
firebase-admin
pytest
trio
anyio